fixed backend for real this time

+++ demo/src/main/java/com/spotify11/demo/repo/PlaylistRepo.java
@@ -10,6 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PlaylistRepo extends JpaRepository<Playlist, Integer> {
+public interface PlaylistRepo extends JpaRepository<Playlist, Long> {
     Playlist findByName(String name);
+    
 }
